+++ k8s/src/main/scala/io/buoyant/k8s/SerializationModule.scala
@@ -1,0 +1,10 @@
+package io.buoyant.k8s
+
+import com.fasterxml.jackson.databind.module.SimpleModule
+
+/**
+ * Class used to provide LoadService-compatible serialization/deserialization plugins to k8s JSON parsing.
+ */
+abstract class SerializationModule {
+  def module: SimpleModule
+}