+++ namerd/iface/control-http/src/main/scala/io/buoyant/namerd/iface/HttpControlService.scala
@@ -1,8 +1,6 @@
 package io.buoyant.namerd.iface
 
-import com.fasterxml.jackson.core.{JsonGenerator, JsonParser}
 import com.fasterxml.jackson.databind._
-import com.fasterxml.jackson.databind.module.SimpleModule
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
 import com.fasterxml.jackson.module.scala.experimental.ScalaObjectMapper
 import com.twitter.finagle.http._
@@ -11,43 +9,16 @@ import com.twitter.finagle.{Status => _, _}
 import com.twitter.io.Buf
 import com.twitter.util._
 import io.buoyant.linkerd.admin.names.{DelegateApiHandler, Delegator}
-import io.buoyant.namerd.DtabStore.{Forbidden, DtabNamespaceDoesNotExistException, DtabVersionMismatchException}
-import io.buoyant.namerd.{DtabStore, Ns, RichActivity, VersionedDtab}
+import io.buoyant.namerd.DtabStore.{DtabNamespaceDoesNotExistException, DtabVersionMismatchException, Forbidden}
+import io.buoyant.namerd.{DtabCodec => DtabModule, DtabStore, Ns, RichActivity, VersionedDtab}
 
 object HttpControlService {
 
   object Json {
-    private[this] def mkModule() = {
-      val module = new SimpleModule
-
-      module.addSerializer(classOf[Path], new JsonSerializer[Path] {
-        override def serialize(path: Path, json: JsonGenerator, p: SerializerProvider): Unit =
-          json.writeString(path.show)
-      })
-      module.addDeserializer(classOf[Path], new JsonDeserializer[Path] {
-        override def deserialize(json: JsonParser, ctx: DeserializationContext) =
-          Path.read(json.getValueAsString)
-      })
-
-      module.addSerializer(classOf[NameTree[Path]], new JsonSerializer[NameTree[Path]] {
-        override def serialize(
-          nameTree: NameTree[Path],
-          json: JsonGenerator,
-          p: SerializerProvider
-        ): Unit = json.writeString(nameTree.show)
-      })
-
-      module.addDeserializer(classOf[NameTree[Path]], new JsonDeserializer[NameTree[Path]] {
-        override def deserialize(json: JsonParser, ctx: DeserializationContext) =
-          NameTree.read(json.getValueAsString)
-      })
-
-      module
-    }
 
     private[this] val mapper = new ObjectMapper with ScalaObjectMapper
     mapper.registerModule(DefaultScalaModule)
-    mapper.registerModule(mkModule())
+    mapper.registerModule(DtabModule.module)
 
     def read[T: Manifest](buf: Buf): Try[T] = {
       val Buf.ByteBuffer.Owned(bb) = Buf.ByteBuffer.coerce(buf)